services/equipment_service/requirements.txt
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.3.1
requests==2.32.3
pika==1.3.2
python-consul==1.1.0
python-dotenv==1.0.1
gunicorn==25.3.0